Moved skin layers into NPCMetaInfo
