+++ core/src/main/java/io/github/scroojalix/npcmanager/commands/subcommands/modifications/SkinModification.java
@@ -40,7 +40,7 @@ public class SkinModification extends SubCommand {
                             }
                         }
                     }
-                    data.getTraits().setSkinLayers(null);
+                    data.getTraits().getMetaInfo().setSkinLayers(null);
                     SkinManager.setSkinFromURL(sender, data, args[4], slimModel);
                     return true;
                 }
@@ -58,7 +58,7 @@ public class SkinModification extends SubCommand {
                                 }
                             }
                         }
-                        data.getTraits().setSkinLayers(null);
+                        data.getTraits().getMetaInfo().setSkinLayers(null);
                         SkinManager.setSkinFromUsername(sender, data, args[4], keepLatest);
                         return true;
                     } else {
@@ -67,7 +67,7 @@ public class SkinModification extends SubCommand {
                 }
             } else if (args[3].equalsIgnoreCase("default")) {
                 data.getTraits().setSkinData(null);
-                data.getTraits().setSkinLayers(null);
+                data.getTraits().getMetaInfo().setSkinLayers(null);
                 main.npc.updateNPC(data);
                 sender.sendMessage(PluginUtils.format("&6Reset the skin of &F" + data.getName()));
                 return true;
